+++ test_data/friedland_xyz_bf_method/earned_premium.csv
@@ -1,0 +1,12 @@
+﻿accident_year,earned_premium,,,,,,,,,,
+1998,"20,000,000",,,,,,,,,,
+1999,"31,500,000",,,,,,,,,,
+2000,"45,000,000",,,,,,,,,,
+2001,"50,000,000",,,,,,,,,,
+2002,"61,183,000",,,,,,,,,,
+2003,"69,175,000",,,,,,,,,,
+2004,"99,322,000",,,,,,,,,,
+2005,"138,151,000",,,,,,,,,,
+2006,"107,578,000",,,,,,,,,,
+2007,"62,438,000",,,,,,,,,,
+2008,"47,797,000",,,,,,,,,,
