+++ test_data/friedland_xyz_bf_method/expected_claim_ratio.csv
@@ -1,0 +1,12 @@
+﻿accident_year,selected_claim_ratio,,,,,,,,,,
+1998,78.3%,,,,,,,,,,
+1999,78.3%,,,,,,,,,,
+2000,78.3%,,,,,,,,,,
+2001,78.3%,,,,,,,,,,
+2002,78.3%,,,,,,,,,,
+2003,78.3%,,,,,,,,,,
+2004,87.1%,,,,,,,,,,
+2005,78.3%,,,,,,,,,,
+2006,65.8%,,,,,,,,,,
+2007,63.8%,,,,,,,,,,
+2008,82.5%,,,,,,,,,,
