+++ test_data/friedland_xyz_bf_method/reported_claims_triangle.csv
@@ -1,0 +1,13 @@
+﻿Accident Year,12,24,36,48,60,72,84,96,108,120,132
+1998,,,"11,171,000","12,380,000","13,216,000","14,067,000","14,688,000","16,366,000","16,163,000","15,835,000","15,822,000"
+1999,,"13,255,000","16,405,000","19,639,000","22,473,000","23,764,000","25,094,000","24,795,000","25,071,000","25,107,000",
+2000,"15,676,000","18,749,000","21,900,000","27,144,000","29,488,000","34,458,000","36,949,000","37,505,000","37,246,000",,
+2001,"11,827,000","16,004,000","21,022,000","26,578,000","34,205,000","37,136,000","38,541,000","38,798,000",,,
+2002,"12,811,000","20,370,000","26,656,000","37,667,000","44,414,000","48,701,000","48,169,000",,,,
+2003,"9,651,000","16,995,000","30,354,000","40,594,000","44,231,000","44,373,000",,,,,
+2004,"16,995,000","40,180,000","58,866,000","71,707,000","70,288,000",,,,,,
+2005,"28,674,000","47,432,000","70,340,000","70,655,000",,,,,,,
+2006,"27,066,000","46,783,000","48,804,000",,,,,,,,
+2007,"19,477,000","31,732,000",,,,,,,,,
+2008,"18,632,000",,,,,,,,,,
+,,,,,,,,,,,
