--- test_data/friedland_xyz_dev_method/current_level_earned_premium.csv
@@ -1,8 +1,0 @@
-﻿Calendar Year,Earned Premiums at 2008 Avg Rate Level,,
-2002," 55,911,228 ",,
-2003," 60,204,386 ",,
-2004," 80,411,091 ",,
-2005," 97,258,304 ",,
-2006," 68,849,920 ",,
-2007," 49,950,400 ",,
-2008," 47,797,000 ",,
