+++ test_framework/__init__.py
@@ -1,0 +1,17 @@
+from test_framework.base import FileWatcher, IDEAutomation, WorkspaceManager, PromptHandler
+from test_framework.harness_base import BaseTestHarness, UnitTestHarness, IntegrationTestHarness
+from test_framework.method_config import MethodConfig, MethodRegistry, StepConfig
+
+__all__ = [
+    'FileWatcher',
+    'IDEAutomation',
+    'WorkspaceManager',
+    'PromptHandler',
+    'BaseTestHarness',
+    'UnitTestHarness',
+    'IntegrationTestHarness',
+    'MethodConfig',
+    'MethodRegistry',
+    'StepConfig',
+]
+
